Add files via upload
Simple implementation of Conway's Game of Life. Allows the user to select an input file, similar to the format of the two example files in this folder. The program will then ask the user for a template output file name, and how many steps the user would like to run. The program then runs the algorithm for each step and writes the current state of the cells at each step to a txt file based on the output file template given by the user.