Add files via upload
Another implementation of Conway's Game of Life. This on is a further extension on the project ConwayGUI, in which the program now uses multiple threads at once to calculate steps in the program, and reduce the grids down to a size that the user's screen can display. I have tested this program to handle up to cell matrices of 50,000 x 50,000 cells. It will then display the resultant final step, or save all steps for you.